refactor(web): rename change product component
From ChangeProductLink to ChangeProductOption
